moved type inference to utils/ instead of core/utils
